Move user settings components to components module
